requirements.txt
SpeechRecognition
elevenlabslib
openai
pyttsx3
sounddevice
soundfile
pyaudio
pyyaml
pyinstaller
pytchat
